refactor(runtime): workers copy the loader vocabulary at spawn
CaptureLoaderVocabulary runs in the WorkerWrapper constructor - on the
parent's JS thread, where the parent's per-isolate state is safely
readable - and the copy rides the wrapper by value into
BackgroundLooper, where InstallLoaderVocabulary writes it into the worker
isolate after runtime init and before any module load. Zero
synchronization: each side only ever touches its own isolate's state.

A worker therefore resolves through the vocabulary its parent had at
spawn; a live worker deliberately does not observe later
reconfiguration - the dev client restarts workers when the vocabulary
changes. The vocabulary types move to the header for the by-value carry.
